Revert "Replace 'unsigned long' with most suitable fixed-size type."
This commit was cherry-picked and was not done, resulting in a few
problems with gcc on 64bit windows.

This reverts commit edd7a72e056b994458ff040d4740b16b35336b60.

Conflicts:
	arch/x86/crc_folding.c
	arch/x86/fill_window_sse.c
	deflate.c
	deflate.h
	match.c
	trees.c

+++ arch/x86/crc_folding.c
@@ -256,7 +256,7 @@ local void partial_fold(deflate_state *const s, const size_t len, __m128i *xmm_c
 }
 
 ZLIB_INTERNAL void crc_fold_copy(deflate_state *const s, unsigned char *dst, const unsigned char *src, long len) {
-    uintptr_t algn_diff;
+    unsigned long algn_diff;
     __m128i xmm_t0, xmm_t1, xmm_t2, xmm_t3;
 
     CRC_LOAD(s)
